Move test script

+++ test.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -x
 #
 # OpenTHC Test Runner
 #
@@ -11,39 +11,77 @@ d=$(dirname "$f")
 
 cd "$d"
 
-output_base="../webroot/test-output"
+output_base="webroot/test-output"
 output_main="$output_base/index.html"
 mkdir -p "$output_base"
 
-search_list="
-../boot.php
-../api/
-../bin/
-../lib/
-../sbin/
-../test/
-../view/
-"
+code_list="boot.php api/ bin/ lib/ sbin/ test/ view/"
 
 
 #
 # Lint
-echo '<h1>Linting...</h1>' > "$output_main"
-find $search_list -type f -name '*.php' -exec php -l {} \; \
-	| grep -v 'No syntax' || true \
-	2>&1 >"$output_base/phplint.txt"
-[ -s "$output_base/phplint.txt" ] || echo "Linting OK" >"$output_base/phplint.txt"
+if [ ! -f "$output_base/phplint.txt" ]
+then
+
+	echo '<h1>Linting...</h1>' > "$output_main"
+
+	find "${code_list[@]}" -type f -name '*.php' -exec php -l {} \; \
+		| grep -v 'No syntax' || true \
+		2>&1 \
+		>"$output_base/phplint.txt"
+
+	[ -s "$output_base/phplint.txt" ] || echo "Linting OK" >"$output_base/phplint.txt"
+
+fi
+
+
+#
+# PHP-CPD
+if [ ! -f "$output_base/phpcpd.txt" ]
+then
+
+	echo '<h1>CPD Check</h1>' > "$output_main"
+
+	vendor/bin/phpcpd \
+		--fuzzy \
+		--log-pmd="$output_base/phpcpd.xml" \
+		--no-ansi \
+		$code_list \
+		> "$output_base/phpcpd.txt"
+
+fi
+
+
+#
+# PHPStan
+if [ ! -f "$output_base/phpstan.html" ]
+then
+
+	echo '<h1>PHPStan...</h1>' > "$output_main"
+
+	vendor/bin/phpstan analyze --error-format=junit --no-progress > "$output_base/phpstan.xml"
+
+	[ -f "phpstan.xsl" ] || curl -qs 'https://openthc.com/pub/phpstan.xsl' > "test/phpstan.xsl"
+
+	xsltproc \
+		--nomkdir \
+		--output "$output_base/phpstan.html" \
+		phpstan.xsl \
+		"$output_base/phpstan.xml"
+
+fi
 
 
 #
 # PHPUnit
 echo '<h1>PHPUnit...</h1>' > "$output_main"
-../vendor/bin/phpunit \
+vendor/bin/phpunit \
 	--verbose \
 	--log-junit "$output_base/output.xml" \
 	--testdox-html "$output_base/testdox.html" \
 	--testdox-text "$output_base/testdox.txt" \
 	--testdox-xml "$output_base/testdox.xml" \
+	test/ \
 	"$@" 2>&1 | tee "$output_base/output.txt"
